+++ src/lib/utils/backup/convertSessions.ts
@@ -6,7 +6,7 @@ import type {
 	ETab,
 	EWindow
 } from '@/lib/types';
-import { compress_options, tabAttr } from '@/lib/constants/env';
+import { compress_options, tabAttr } from '@/lib/constants';
 import { compress as compressLZ } from 'lz-string';
 import { compress } from '..';
