+++ src/lib/utils/log.ts
@@ -1,4 +1,4 @@
 import pino from 'pino';
-import { isDEV } from '@constants/env';
+import { isDEV } from '@constants/shared';
 
 export const log = pino({ level: isDEV ? 'debug' : 'warn' });
